chore: fix clippy 1.95 lints (CI toolchain upgrade)
- checked_div instead of manual zero-guard (lru_index.rs)
- zip-based counter loops instead of explicit entry_offset (FUSE readdir)
- remove unnecessary usize-to-usize cast (coordinator core)
- sort_by_key instead of sort_by with single-field cmp (tile_progress)

+++ xearthlayer/src/prefetch/adaptive/coordinator/core.rs
@@ -969,7 +969,7 @@ impl AdaptivePrefetchCoordinator {
         // Update statistics
         self.total_cycles += 1;
         self.total_tiles_submitted += submitted as u64;
-        self.total_cache_hits += (plan.skipped_cached as usize + total_filtered) as u64;
+        self.total_cache_hits += (plan.skipped_cached + total_filtered) as u64;
 
         // Update shared status for TUI
         self.update_shared_status(position, &plan, submitted);
